+++ src/reduct/core/common.h
@@ -1,0 +1,15 @@
+// Copyright 2023 ReductStore
+// This Source Code Form is subject to the terms of the Mozilla Public
+//    License, v. 2.0. If a copy of the MPL was not distributed with this
+//    file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+#ifndef REDUCTSTORE_COMMON_H
+#define REDUCTSTORE_COMMON_H
+
+#include <string>
+
+namespace reduct::core {
+constexpr static std::string_view kContentTypeOctetStream = "application/octet-stream";
+}  // namespace reduct::core
+
+#endif  // REDUCTSTORE_COMMON_H